+++ Model_30M/models.py
@@ -20,7 +20,7 @@ class CONFIG_30M(models.Model):
     quantity                = models.BigIntegerField(verbose_name='QUANTITY',default=0)
     return_price            = models.FloatField(verbose_name='RET_PRICE',default=0)
     ltp                     = models.FloatField(verbose_name='LTP',default=0)
-    count                   = models.BigIntegerField(verbose_name='COUNT',default=0)
+    count                   = models.FloatField(verbose_name='COUNT',default=0)
     order_id                = models.BigIntegerField(verbose_name='ORDER ID',default=0)
     order_status            = models.CharField(max_length=100, verbose_name='ORDER STATUS',default='NONE')
     class Meta:
